WINDUPRULE-226 Rule for detecting Hibernate and JSF embedded library: splitted rules
